+++ src/components/app/app.ts
@@ -3,11 +3,11 @@ import { Title, DOCUMENT } from '@angular/platform-browser';
 
 import { IonicApp } from './app-root';
 import * as Constants from './app-constants';
-import { ClickBlock } from '../click-block/click-block';
+import { ClickBlock } from './click-block';
 import { runInDev } from '../../util/util';
 import { Config } from '../../config/config';
 import { isNav, NavOptions, DIRECTION_FORWARD, DIRECTION_BACK } from '../../navigation/nav-util';
-import { MenuController } from '../menu/menu-controller';
+import { MenuController } from './menu-controller';
 import { NavController } from '../../navigation/nav-controller';
 import { Platform } from '../../platform/platform';
 import { ViewController } from '../../navigation/view-controller';