+++ src/components/app/menu-controller.ts
@@ -1,5 +1,4 @@
-import { Menu } from './menu';
-import { MenuType } from './menu-types';
+import { Menu, MenuType } from './menu-interface';
 import { Platform } from '../../platform/platform';
 import { removeArrayItem, assert } from '../../util/util';
 
@@ -249,7 +248,9 @@ export class MenuController {
       // there could be more than one menu on the same side
       // so first try to get the enabled one
       menu = this._menus.find(m => m.side === menuId && m.enabled);
-      if (menu) return menu;
+      if (menu) {
+        return menu;
+      }
 
       // didn't find a menu side that is enabled
       // so try to get the first menu side found
@@ -343,3 +344,5 @@ export class MenuController {
 }
 
 let menuTypes: { [name: string]: new(...args: any[]) => MenuType } = {};
+
+