+++ src/components/tabs/test/advanced/pages/tab1-page1/tab1-page1.ts
@@ -28,9 +28,7 @@ export class Tab1Page1 {
   }
 
   logout() {
-    this.app.getRootNav().setRoot('sign-in', null, { animate: true, direction: 'back' }).catch(() => {
-      console.debug('logout cancelled');
-    });
+    this.app.getRootNav().setRoot('sign-in', null, { animate: true, direction: 'back' });
   }
 
   ionViewWillEnter() {