+++ src/components/toggle/toggle.ts
@@ -1,4 +1,4 @@
-import { AfterViewInit, ChangeDetectorRef, Component, ElementRef, forwardRef, HostListener, Input, OnDestroy, Optional, Renderer, ViewEncapsulation } from '@angular/core';
+import { NgZone, AfterViewInit, ChangeDetectorRef, Component, ElementRef, HostListener, Input, OnDestroy, Optional, Renderer, ViewEncapsulation } from '@angular/core';
 import { NG_VALUE_ACCESSOR } from '@angular/forms';
 
 import { Config } from '../../config/config';
@@ -12,13 +12,6 @@ import { Item } from '../item/item';
 import { KEY_ENTER, KEY_SPACE } from '../../platform/key';
 import { Platform } from '../../platform/platform';
 import { ToggleGesture } from './toggle-gesture';
-
-
-export const TOGGLE_VALUE_ACCESSOR: any = {
-  provide: NG_VALUE_ACCESSOR,
-  useExisting: forwardRef(() => Toggle),
-  multi: true
-};
 
 /**
  * @name Toggle
@@ -60,7 +53,7 @@ export const TOGGLE_VALUE_ACCESSOR: any = {
 @Component({
   selector: 'ion-toggle',
   template:
-    '<div class="toggle-icon" [class.toggle-checked]="_value" [class.toggle-activated]="_activated">' +
+    '<div class="toggle-icon">' +
       '<div class="toggle-inner"></div>' +
     '</div>' +
     '<button role="checkbox" ' +
@@ -70,12 +63,14 @@ export const TOGGLE_VALUE_ACCESSOR: any = {
             '[attr.aria-checked]="_value" ' +
             '[attr.aria-labelledby]="_labelId" ' +
             '[attr.aria-disabled]="_disabled" ' +
-            'class="item-cover">' +
+            'class="item-cover" disable-activated>' +
     '</button>',
   host: {
-    '[class.toggle-disabled]': '_disabled'
+    '[class.toggle-disabled]': '_disabled',
+    '[class.toggle-checked]': '_value',
+    '[class.toggle-activated]': '_activated',
   },
-  providers: [TOGGLE_VALUE_ACCESSOR],
+  providers: [ { provide: NG_VALUE_ACCESSOR, useExisting: Toggle, multi: true } ],
   encapsulation: ViewEncapsulation.None,
 })
 export class Toggle extends BaseInput<boolean> implements IonicTapInput, AfterViewInit, OnDestroy  {
@@ -107,7 +102,8 @@ export class Toggle extends BaseInput<boolean> implements IonicTapInput, AfterVi
     @Optional() item: Item,
     private _gestureCtrl: GestureController,
     private _domCtrl: DomController,
-    private _cd: ChangeDetectorRef
+    private _cd: ChangeDetectorRef,
+    private _zone: NgZone,
   ) {
     super(config, elementRef, renderer, 'toggle', false, form, item, null);
   }
@@ -124,16 +120,13 @@ export class Toggle extends BaseInput<boolean> implements IonicTapInput, AfterVi
   /**
    * @hidden
    */
-  _inputNormalize(val: any): boolean {
-    return isTrueProperty(val);
-  }
+  _inputCheckHasValue() {}
 
   /**
    * @hidden
    */
-  _inputUpdated() {
-    this._item && this._item.setElementClass('item-toggle-checked', this.value);
-    this._cd.detectChanges();
+  _inputNormalize(val: any): boolean {
+    return isTrueProperty(val);
   }
 
   /**
@@ -143,9 +136,11 @@ export class Toggle extends BaseInput<boolean> implements IonicTapInput, AfterVi
     assert(startX, 'startX must be valid');
     console.debug('toggle, _onDragStart', startX);
 
-    this._startX = startX;
-    this._fireFocus();
-    this._activated = true;
+    this._zone.run(() => {
+      this._startX = startX;
+      this._fireFocus();
+      this._activated = true;
+    });
   }
 
   /**
@@ -157,22 +152,32 @@ export class Toggle extends BaseInput<boolean> implements IonicTapInput, AfterVi
       return;
     }
 
-    console.debug('toggle, _onDragMove', currentX);
+    let dirty = false;
+    let value: boolean;
+    let activated: boolean;
 
     if (this._value) {
       if (currentX + 15 < this._startX) {
-        this.value = false;
-        this._haptic.selection();
-        this._startX = currentX;
-        this._activated = true;
+        dirty = true;
+        value = false;
+        activated = true;
       }
 
     } else if (currentX - 15 > this._startX) {
-      this.value = true;
-      this._haptic.selection();
-      this._startX = currentX;
-      this._activated = (currentX < this._startX + 5);
+      dirty = true;
+      value = true;
+      activated = (currentX < this._startX + 5);
     }
+
+    if (dirty) {
+      this._zone.run(() => {
+        this.value = value;
+        this._startX = currentX;
+        this._activated = activated;
+        this._haptic.selection();
+      });
+    }
+
   }
 
   /**
@@ -185,20 +190,22 @@ export class Toggle extends BaseInput<boolean> implements IonicTapInput, AfterVi
     }
     console.debug('toggle, _onDragEnd', endX);
 
-    if (this._value) {
-      if (this._startX + 4 > endX) {
-        this.value = false;
+    this._zone.run(() => {
+      if (this._value) {
+        if (this._startX + 4 > endX) {
+          this.value = false;
+          this._haptic.selection();
+        }
+
+      } else if (this._startX - 4 < endX) {
+        this.value = true;
         this._haptic.selection();
       }
 
-    } else if (this._startX - 4 < endX) {
-      this.value = true;
-      this._haptic.selection();
-    }
-
-    this._activated = false;
-    this._fireBlur();
-    this._startX = null;
+      this._activated = false;
+      this._fireBlur();
+      this._startX = null;
+    });
   }
 
   /**