+++ src/components/toolbar/toolbar-item.ts
@@ -3,7 +3,7 @@ import { ContentChildren, Directive, ElementRef, forwardRef, Optional, Inject, R
 import { Button } from '../button/button';
 import { Config } from '../../config/config';
 import { Ion } from '../ion';
-import { Navbar } from '../navbar/navbar';
+import { Navbar } from './navbar';
 import { Toolbar } from './toolbar';