+++ src/platform/platform-registry.ts
@@ -1,7 +1,7 @@
 import { OpaqueToken } from '@angular/core';
 
 import { Platform, PlatformConfig } from './platform';
-import { isCordova, isIos, isIosUIWebView } from './platform-utils';
+import { isCordova, isElectron, isIos, isIosUIWebView } from './platform-utils';
 
 
 export const PLATFORM_CONFIGS: { [key: string]: PlatformConfig } = {
@@ -219,6 +219,24 @@ export const PLATFORM_CONFIGS: { [key: string]: PlatformConfig } = {
     isMatch(plt: Platform): boolean {
       return isCordova(plt);
     }
+  },
+
+  /**
+   * electron
+   */
+  'electron': {
+    superset: 'core',
+    initialize: function(plt: Platform) {
+      plt.prepareReady = function() {
+        // 1) ionic bootstrapped
+        plt.windowLoad(function() {
+          plt.triggerReady('electron');
+        });
+      };
+    },
+    isMatch(plt: Platform): boolean {
+      return isElectron(plt);
+    }
   }
 };
 
@@ -228,3 +246,4 @@ export const PlatformConfigToken = new OpaqueToken('PLTCONFIG');
 export function providePlatformConfigs() {
   return PLATFORM_CONFIGS;
 }
+